Pin Python version to 3.11.x
Why pin Python version?

Because we _can_ put everybody on the same version.
This way,
we make it less likely different people will hit different issues.

Why use Python 3.11?

It is well established,
binary wheels exist for it in PyPI,
it is available on UBI 8 and 9, and on Fedora starting with Fedora 37.

Why say `python = ">=3.11.5, <3.12.0"`?

The syntax uses basic math instead of tildes and carets,
which not everybody might've learned what they mean.
That makes the requirement easier to read and understand.

Signed-off-by: Jiri Daněk <jdanek@redhat.com>